Create TASK1
The user can predict a randomly generated number inside a given range with this C++ program. It asks for the bottom and upper boundaries of the range, uses `rand()` to produce a random number inside the range, and then asks the user to estimate the number many times. After each guess, it gives output ("Too low!" or "Too high!") if the number is not rightly guessed and, congratulations the user, and ends the game when the user guesses the right number.